Add megapacker.pl script to create .mgu updater files
This is a relatively simple Perl script that takes app and/or updater
ESP32 .bin image(s) and packs them into the .mgu format that is expected
by the MegaGRRL firmware when initiating firmware updates from the File
Browser.

Example usage, for an app-only update on the (d)esktop version:

	perl utils/megapacker.pl --app=firmware/build/megagrrl.bin --hw=d --outfile=megagrrl.mgu

To update both the app and updater images for the (p)ortable version:

	perl utils/megapacker.pl --app=megagrrl_app.bin --updater=megagrrl_updater.bin --hw=p --outfile=megagrrl_portable.mgu

To specify the version rather than extract it from the header in the
.bin file (note: the script defaults to hw version 'd'):

	perl utils/megapacker.pl --app=megagrrl.bin --version=v0.97dev --outfile=megagrrl_v0.97dev.mgu

No additional dependencies should be required if using any non-ancient
Perl.